Moved googleSheetsService.js file to new Services folder
